+++ lib/torchvision/datasets/image_folder.rb
@@ -1,8 +1,10 @@
 module TorchVision
   module Datasets
     class ImageFolder < DatasetFolder
-      def initialize(root, transform: nil)
-        super
+      IMG_EXTENSIONS = [".jpg", ".jpeg", ".png", ".ppm", ".bmp", ".pgm", ".tif", ".tiff", ".webp"]
+
+      def initialize(root, transform: nil, target_transform: nil, is_valid_file: nil)
+        super(root, extensions: IMG_EXTENSIONS, transform: transform, target_transform: target_transform, is_valid_file: is_valid_file)
         @imgs = @samples
       end
     end