+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.2.1 (unreleased)
+## 0.2.1 (2021-03-14)
 
 - Added `ImageFolder` and `DatasetFolder`
 - Added `CenterCrop` and `RandomResizedCrop` transforms